+++ lib/desertcart/error.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module LedgerSync
+  class Error
+    class OperationError
+      class UnprocessableEntityError < OperationError; end
+    end
+  end
+end